Update touch drivers

+++ lib/XPT2046_Touchscreen_ID542/XPT2046_Touchscreen.h
@@ -47,7 +47,7 @@ public:
 	TS_Point getPoint();
 	bool tirqTouched();
 	bool touched();
-	void readData(uint16_t *x, uint16_t *y, uint8_t *z);
+	void readData(int16_t *x, int16_t *y, int16_t *z);
 	bool bufferEmpty();
 	uint8_t bufferSize() { return 1; }
 	void setRotation(uint8_t n) { rotation = n % 4; }
